# (WIP) MINOR: Cache topic resolution in TopicIds set