@dajac 2024-09-26
Replaced by https://github.com/apache/kafka/pull/17285.